move onbounded stake checks to voting module
