+++ dao-contracts/tests/features/bid_escrow/external_not_to_va_worker_failed.feature
@@ -47,8 +47,8 @@ Feature: External Worker who doesn't want to become a VA submits job
       | InternalWorker   | 0            | 1000         | 0          |
       | VA1              | 0            | 1000         | 500        |
       | VA2              | 0            | 1000         | 500        |
-    And total unbounded stake for voting 0 is 50 tokens
-    And ballot for voting 0 for ExternalWorker has 50 unbounded tokens
+    And BidEscrow total unbounded stake for voting 0 is 50 tokens
+    And BidEscrow ballot for voting 0 for ExternalWorker has 50 unbounded tokens
     When 6 days passed
     And informal voting with id 0 ends in BidEscrow contract
     Then balances are
@@ -60,8 +60,8 @@ Feature: External Worker who doesn't want to become a VA submits job
       | InternalWorker   | 0            | 1000         | 0          |
       | VA1              | 0            | 1000         | 0          |
       | VA2              | 0            | 1000         | 0          |
-    And total unbounded stake for voting 0 is 50 tokens
-    And ballot for voting 0 for ExternalWorker has 50 unbounded tokens
+    And BidEscrow total unbounded stake for voting 0 is 50 tokens
+    And BidEscrow ballot for voting 0 for ExternalWorker has 50 unbounded tokens
     When voters vote in BidEscrow formal voting with id 0
       | account          | REP stake | choice |
      #| ExternalWorker   | 50        | Yes    | - automatically voted by the system
