+++ dao-contracts/tests/features/bid_escrow/onboarding_request_failure.feature
@@ -35,8 +35,8 @@ Feature: User who wants to become a VA without submitting a job
       | Bob              | 1000         | 0            | 0          |
       | VA1              | 0            | 1000         | 500        |
       | VA2              | 0            | 1000         | 500        |
-    And total onboarding unbounded stake for voting 0 is 100 tokens
-    And ballot for onboarding voting 0 for Bob has 100 unbounded tokens
+    And Onboarding total unbounded stake for voting 0 is 100 tokens
+    And Onboarding ballot for voting 0 for Bob has 100 unbounded tokens
     When 6 days passed
     And informal voting with id 0 ends in Onboarding contract
     Then balances are
@@ -46,8 +46,8 @@ Feature: User who wants to become a VA without submitting a job
       | Bob              | 1000         | 0            | 0          |
       | VA1              | 0            | 1000         | 0          |
       | VA2              | 0            | 1000         | 0          |
-    And total onboarding unbounded stake for voting 0 is 100 tokens
-    And ballot for onboarding voting 0 for Bob has 100 unbounded tokens
+    And Onboarding total unbounded stake for voting 0 is 100 tokens
+    And Onboarding ballot for voting 0 for Bob has 100 unbounded tokens
     When voters vote in Onboarding formal voting with id 0
       | user    | REP stake  | choice | 
      #| Bob     | 100        | yes    | - automatically voted by the system - 1000CSPR converted to 100 Reputation
