+++ dao-contracts/tests/features/variables/voting_clearness_delta.feature
@@ -34,7 +34,7 @@ Feature: Voting clearness delta
       | BidEscrow | 0 | formal |
   Scenario: Results are close - time between votings is doubled
     When InternalWorker submits the JobProof of Job 0
-    And voters vote in BidEscrow formal voting with id 0
+    And voters vote in BidEscrow informal voting with id 0
       | account          | REP stake | choice |
      #| InternalWorker   | 100       | Yes    | - automatically voted by the system
       | VA1              | 450       | Yes    |
